requirements.txt
streamlit
langchain
langchain-google-genai
python-dotenv
streamlit-extras
speechrecognition
pydub
openai-whisper
pyaudio